+++ src/unnest.h
@@ -60,15 +60,15 @@ void add_node(UT& U, AT& acc, VarAccumulator& vacc,
 
 struct Unnester {
 
-  enum Unnamed {NONE, STACK, EXCLUDE};
-  Unnamed sexp2unnamed(SEXP x);
+  enum ProcessUnnamed {NONE, STACK, EXCLUDE, ASIS, PASTE};
+  ProcessUnnamed sexp2unnamed(SEXP x);
 
 
   bool dedupe;
   bool stack_atomic;
   bool stack_atomic_df;
   Spec::Process process_atomic;
-  Unnamed unnamed_list;
+  ProcessUnnamed process_unnamed_list;
   bool rep_to_max;
 
   cpair2ix_map cp2i;
@@ -171,28 +171,35 @@ struct Unnester {
     // LENGTH(X) > 0 and X != NULL in here
     if (pspec.process == Spec::Process::ASIS) {
       acc.pnodes.push_front(make_unique<AsIsNode>(ix, x));
-      P("<--- added ASIS node impl:%s(%ld) acc[%ld,%ld]\n",
-        full_name(ix).c_str(), ix, acc.nrows, acc.pnodes.size());
+      P("<--- added ASIS node impl:%s(%ld) acc[%ld,%ld]\n", full_name(ix).c_str(), ix, acc.nrows, acc.pnodes.size());
     } else if (pspec.process == Spec::Process::PASTE) {
       acc.pnodes.push_front(make_unique<PasteNode>(ix, x));
-      P("<--- added PASTE node impl:%s(%ld) acc[%ld,%ld]\n",
-        full_name(ix).c_str(), ix, acc.nrows, acc.pnodes.size());
+      P("<--- added PASTE node impl:%s(%ld) acc[%ld,%ld]\n", full_name(ix).c_str(), ix, acc.nrows, acc.pnodes.size());
     } else if (TYPEOF(x) == VECSXP) {
       // Lists
-      if (this->unnamed_list == Unnamed::EXCLUDE &&
-          pspec.terminal && spec.stack == Spec::Stack::AUTO &&
-          Rf_getAttrib(x, R_NamesSymbol) == R_NilValue) {
+      bool is_unnamed = Rf_getAttrib(x, R_NamesSymbol) == R_NilValue;
+      if (is_unnamed && pspec.terminal && spec.stack == Spec::Stack::AUTO) {
         /* PP("pspec: %s\n", pspec.to_string().c_str()); */
         /* PP("spec: %s\n", spec.to_string().c_str()); */
-        return;
+        if (this->process_unnamed_list == ProcessUnnamed::EXCLUDE) {
+          P("<--- excluded UNNAMED node impl:%s(%ld) acc[%ld,%ld]\n", full_name(ix).c_str(), ix, acc.nrows, acc.pnodes.size());
+          return;
+        } else if (this->process_unnamed_list == ProcessUnnamed::ASIS) {
+          acc.pnodes.push_front(make_unique<AsIsNode>(ix, x));
+          P("<--- added UNNAMED-ASIS node impl:%s(%ld) acc[%ld,%ld]\n", full_name(ix).c_str(), ix, acc.nrows, acc.pnodes.size());
+          return;
+        } else if (this->process_unnamed_list == ProcessUnnamed::PASTE) {
+          acc.pnodes.push_front(make_unique<PasteNode>(ix, x));
+          P("<--- added UNNAMED-PASTE node impl:%s(%ld) acc[%ld,%ld]\n", full_name(ix).c_str(), ix, acc.nrows, acc.pnodes.size());
+          return;
+        }
       }
       stack_atomic = stack_atomic || this->stack_atomic_df && is_data_frame(x);
       P("--> add_node_impl:%s(%ld) %s\n", full_name(ix).c_str(), ix, spec.to_string().c_str());
       const vector<SpecMatch>& matches = spec.match(x);
       P("    nr. matches: %ld, stack_atomic: %d\n", matches.size(), stack_atomic);
       if (spec.stack == Spec::Stack::STACK ||
-          (this->unnamed_list == Unnamed::STACK &&
-           Rf_getAttrib(x, R_NamesSymbol) == R_NilValue)) {
+          (is_unnamed && this->process_unnamed_list == ProcessUnnamed::STACK)) {
         stack_nodes(acc, vacc, pspec, spec, ix, matches, stack_atomic);
       } else {
         spread_nodes(acc, vacc, pspec, spec, ix, matches, stack_atomic);