+++ NEWS.md
@@ -3,6 +3,7 @@ Version 0.0.2.9000 [Dev]
 
 - New `default` argument in `s()` specification.
 - Data frame columns are now stacked by default (`stack_atomic = NULL` parameter)
+- New `unnest()` parameter `process_unnamed_lists` for controlling processing of unnamed lists.
 
 Version 0.0.2
 =============